moves & updates resources
