--- src/Filament/OdkAdmin/Resources/DatasetResource/Pages/ViewDataset.php
@@ -1,32 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\DatasetResource\Pages;
-
-use Filament\Resources\Pages\ViewRecord;
-use Filament\Resources\Resource;
-use Illuminate\Contracts\Support\Htmlable;
-use Illuminate\Database\Eloquent\Model;
-use phpDocumentor\Reflection\Types\ClassString;
-use Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\DatasetResource;
-use Stats4sd\FilamentOdkLink\Models\OdkLink\Dataset;
-
-class ViewDataset extends ViewRecord
-{
-    protected static string $resource = DatasetResource::class;
-
-    /**
-     * @phpstan-return Dataset
-     */
-    public function getRecord(): Model | Dataset
-    {
-        /** @var Dataset $record */
-        $record = parent::getRecord();
-
-        return $record;
-    }
-
-    public function getTitle(): string | Htmlable
-    {
-        return 'Dataset: ' . $this->getRecord()->name;
-    }
-}
