--- src/Filament/OdkAdmin/Resources/DatasetResource/RelationManagers/VariablesRelationManager.php
@@ -1,63 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\DatasetResource\RelationManagers;
-
-use Filament\Forms;
-use Filament\Forms\Components\TextInput;
-use Filament\Forms\Form;
-use Filament\Resources\RelationManagers\RelationManager;
-use Filament\Tables;
-use Filament\Tables\Columns\TextColumn;
-use Filament\Tables\Table;
-
-class VariablesRelationManager extends RelationManager
-{
-    protected static string $relationship = 'variables';
-
-    public function isReadOnly(): bool
-    {
-        return false;
-    }
-
-    public function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                TextInput::make('name')
-            ->label('Enter the variable name as it should appear in a dataset (column header)')
-            ->helperText('Ideally, this should be in snake_case (e.g. "productive_activities")'),
-                Forms\Components\TextInput::make('label')
-                    ->label('The label for the variable')
-                    ->required()
-                    ->maxLength(255),
-                Forms\Components\Textarea::make('description')
-            ])
-            ->columns(1);
-    }
-
-    public function table(Table $table): Table
-    {
-        return $table
-            ->recordTitleAttribute('name')
-            ->columns([
-                Tables\Columns\TextColumn::make('id'),
-                Tables\Columns\TextColumn::make('name'),
-                TextColumn::make('label'),
-            ])
-            ->filters([
-                //
-            ])
-            ->headerActions([
-                Tables\Actions\CreateAction::make(),
-            ])
-            ->actions([
-                Tables\Actions\EditAction::make(),
-                Tables\Actions\DeleteAction::make(),
-            ])
-            ->bulkActions([
-                Tables\Actions\BulkActionGroup::make([
-                    Tables\Actions\DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-}
