--- src/Filament/OdkAdmin/Resources/DatasetResource/RelationManagers/XlsformTemplatesRelationManager.php
@@ -1,56 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\DatasetResource\RelationManagers;
-
-use Filament\Forms;
-use Filament\Forms\Form;
-use Filament\Resources\RelationManagers\RelationManager;
-use Filament\Tables;
-use Filament\Tables\Table;
-use Illuminate\Database\Eloquent\Model;
-
-class XlsformTemplatesRelationManager extends RelationManager
-{
-    protected static string $relationship = 'xlsformTemplates';
-
-    public static function getTitle(Model $ownerRecord, string $pageClass): string
-    {
-        return 'Used in... ';
-    }
-
-    public function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                Forms\Components\TextInput::make('title')
-                    ->required()
-                    ->maxLength(255),
-            ]);
-    }
-
-    public function table(Table $table): Table
-    {
-        return $table
-            ->recordTitleAttribute('title')
-            ->columns([
-                Tables\Columns\TextColumn::make('title')
-                    ->label('Xlsform Template'),
-                Tables\Columns\TextColumn::make('active_xlsforms_count')->counts('xlsforms')
-                    ->label('# of Active Xlsforms')])
-            ->filters([
-                //
-            ])
-            ->headerActions([
-                Tables\Actions\CreateAction::make(),
-            ])
-            ->actions([
-                Tables\Actions\EditAction::make(),
-                Tables\Actions\DeleteAction::make(),
-            ])
-            ->bulkActions([
-                Tables\Actions\BulkActionGroup::make([
-                    Tables\Actions\DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-}
