+++ src/Filament/OdkAdmin/Resources/Datasets/Pages/EditDataset.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\Datasets\Pages;
+
+use Filament\Actions;
+use Illuminate\Database\Eloquent\Model;
+use Filament\Resources\Pages\EditRecord;
+use Stats4sd\FilamentOdkLink\Models\OdkLink\Dataset;
+use Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\Datasets\DatasetResource;
+
+class EditDataset extends EditRecord
+{
+    protected static string $resource = DatasetResource::class;
+
+    /**
+     * @phpstan-return Dataset
+     */
+    public function getRecord(): Model | Dataset
+    {
+        /** @var Dataset $record */
+        $record = parent::getRecord();
+
+        return $record;
+    }
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\DeleteAction::make(),
+        ];
+    }
+}
