+++ src/Filament/OdkAdmin/Resources/Datasets/Schemas/DatasetInfoList.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\XlsformTemplates\Schemas;
+
+use Filament\Schemas\Schema;
+use Filament\Schemas\Components\Section;
+use Filament\Infolists\Components\TextEntry;
+
+class DatasetInfoList
+{
+    public static function configure(Schema $schema): Schema
+    {
+        return $schema
+            ->schema([
+               Section::make('Dataset Details')
+                    ->schema([
+                        TextEntry::make('name'),
+                        TextEntry::make('primary_key'),
+                        TextEntry::make('description'),
+                    ])
+                    ->columns([
+                        'lg' => 3,
+                        'md' => 2,
+                        'sm' => 1,
+                    ]),
+            ]);
+    }
+}
