+++ src/Filament/OdkAdmin/Resources/Datasets/Schemas/XlsformTemplatesRelationManagerForm.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\Datasets\Schemas;
+
+use Filament\Schemas\Schema;
+use Filament\Forms\Components\TextInput;
+
+class XlsformTemplatesRelationManagerForm
+{
+    public static function configure(Schema $schema): Schema
+    {
+        return $schema
+            ->schema([
+                TextInput::make('title')
+                    ->required()
+                    ->maxLength(255),
+            ]);
+    }
+}
