--- src/Filament/OdkAdmin/Resources/XlsformModuleResource.php
@@ -1,75 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources;
-
-use Filament\Forms;
-use Filament\Forms\Form;
-use Filament\Resources\Resource;
-use Filament\Tables;
-use Filament\Tables\Table;
-use Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\XlsformModuleResource\Pages\ManageXlsformModule;
-use Stats4sd\FilamentOdkLink\Models\OdkLink\XlsformModule;
-
-class XlsformModuleResource extends Resource
-{
-    protected static ?string $model = XlsformModule::class;
-
-    protected static ?string $navigationIcon = 'heroicon-o-rectangle-stack';
-
-    protected static ?string $navigationGroup = 'ODK Forms and Datasets';
-
-
-    public static function form(Form $form): Form
-    {
-        return $form
-            ->columns(1)
-            ->schema([
-                Forms\Components\TextInput::make('label')
-                    ->label('Enter the readable name of the module type')
-                    ->hint('e.g. "Dietary Diversity"')
-                    ->required()
-                    ->maxLength(255),
-                Forms\Components\TextInput::make('name')
-                    ->label('Enter the name of the group')
-                    ->hint('e.g. "dietary_diversity"')
-                    ->required()
-                    ->maxLength(255),
-            ]);
-    }
-
-    public static function table(Table $table): Table
-    {
-        return $table
-            ->columns([
-                Tables\Columns\TextColumn::make('label')
-                    ->searchable(),
-                Tables\Columns\TextColumn::make('name')
-                    ->searchable(),
-            ])
-            ->filters([
-                //
-            ])
-            ->actions([
-                Tables\Actions\EditAction::make(),
-            ])
-            ->bulkActions([
-                Tables\Actions\BulkActionGroup::make([
-                    Tables\Actions\DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-
-    public static function getRelations(): array
-    {
-        return [
-            //
-        ];
-    }
-
-    public static function getPages(): array
-    {
-        return [
-            'index' => ManageXlsformModule::route('/'),
-        ];
-    }
-}
