--- src/Filament/OdkAdmin/Resources/XlsformModuleResource/Pages/ManageXlsformModule.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\XlsformModuleResource\Pages;
-
-use Filament\Actions;
-use Filament\Resources\Pages\ListRecords;
-use Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\XlsformModuleResource;
-
-class ManageXlsformModule extends ListRecords
-{
-    protected static string $resource = XlsformModuleResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\CreateAction::make(),
-        ];
-    }
-}
