--- src/Filament/OdkAdmin/Resources/XlsformModuleVersionResource/Pages/ManageXlsformModuleVersion.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\XlsformModuleVersionResource\Pages;
-
-use Filament\Actions;
-use Filament\Resources\Pages\ManageRecords;
-use Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\XlsformModuleVersionResource;
-
-class ManageXlsformModuleVersion extends ManageRecords
-{
-    protected static string $resource = XlsformModuleVersionResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\CreateAction::make(),
-        ];
-    }
-}
