--- src/Filament/OdkAdmin/Resources/XlsformTemplateResource/RelationManagers/XlsformModuleRelationManager.php
@@ -1,56 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\XlsformTemplateResource\RelationManagers;
-
-use Filament\Forms;
-use Filament\Forms\Form;
-use Filament\Resources\RelationManagers\RelationManager;
-use Filament\Tables;
-use Filament\Tables\Table;
-
-class XlsformModuleRelationManager extends RelationManager
-{
-    protected static string $relationship = 'xlsformModules';
-
-    public function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                Forms\Components\TextInput::make('label')
-                    ->required()
-                    ->maxLength(255),
-            ]);
-    }
-
-    public function table(Table $table): Table
-    {
-        return $table
-            ->recordTitleAttribute('label')
-            ->columns([
-                Tables\Columns\TextColumn::make('label'),
-                Tables\Columns\TextColumn::make('default_survey_rows_count')
-                    ->label('# Survey Rows')
-                    ->counts('defaultSurveyRows'),
-                Tables\Columns\TextColumn::make('default_choice_lists_count')
-                    ->label('# Choice Listes')
-                    ->counts('defaultChoiceLists'),
-                Tables\Columns\TextColumn::make('defaultLocales.language_label')
-                    ->label('Available Locales'),
-            ])
-            ->filters([
-                //
-            ])
-            ->headerActions([
-                Tables\Actions\CreateAction::make(),
-            ])
-            ->actions([
-                Tables\Actions\EditAction::make(),
-                Tables\Actions\DeleteAction::make(),
-            ])
-            ->bulkActions([
-                Tables\Actions\BulkActionGroup::make([
-                    Tables\Actions\DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-}
